Move the reference builder into the test module, and the packed code onto Operation
The one-pass builder existed only as the byte-identity oracle for the
counting sort's equivalence tests, but sat cfg(test)-gated in the body
of builder.rs and key_segment.rs. It now lives inside the test module
that runs those tests, self-contained, and the production files carry
only the counting sort: builder.rs loses BuilderKey and BuilderKeyLists,
key_segment.rs loses KeySegment::build.

The 2-bit operation code the packed references carry moves onto
Operation as packed_code/from_packed_code, documented as distinct from
the serialization wire values, which pin a different numbering.
